+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Unreleased
+- Add DynamicContractObject: which is created from DefiWalletActor
+- Add DynamicContractObject::Send : for modifying contract
+- Add DynamicContractObject::Call : for query contract
+- Add DynamicContractObject::Encode : for wallet connect
 - Support Unreal Engine 5.1.0
 - play-cpp-sdk v0.0.13-alpha
 - Save session to a sessioninfo.json file under project dir